pyproject.toml
[project]
name = "lucerna"
version = "0.1.0"
description = ""
authors = [
    {name = "qwerltaz"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gitpython (>=3.1.45,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
